+++ custom_components/gios/const.py
@@ -1,6 +1,12 @@
 """Constants for GIOS integration."""
+ATTR_AQI = "AQI"
 ATTR_ID = "id"
+ATTR_INDEX = "index"
+ATTR_NAME = "name"
+ATTR_STATION = "station"
+ATTR_VALUE = "value"
 CONF_STATION_ID = "station_id"
+DATA_CLIENT = "client"
 DEFAULT_SCAN_INTERVAL = 1800
 DEFAULT_NAME = "GIOŚ"
 DOMAIN = "gios"
